requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
docker-pycreds==0.4.0
dominate==2.7.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
jsonpatch==1.32
jsonpointer==2.3
magic-filter==1.0.9
multidict==6.0.4
networkx==3.0
numpy==1.24.1
pathtools==0.1.2
Pillow==9.4.0
protobuf==4.21.12
psutil==5.9.4
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
scipy==1.10.0
sentry-sdk==1.13.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
torch==1.13.1+cpu
torchaudio==0.13.1+cpu
torchvision==0.14.1+cpu
tornado==6.2
typing_extensions==4.4.0
urllib3==1.26.14
visdom==0.2.3
wandb==0.13.9
websocket-client==1.4.2
yarl==1.8.2
